requirements.txt
starlette==0.13.4

aiofiles==0.5.0  # StaticFiles support
jinja2==2.11.2
libsass==0.20.0  # Sass compilation support
react==4.3.0  # Server side rendered React.js
